# api/app/docs/swagger.yaml
definitions:
...
  hrms_modules_worklog_internal_dto.FTItem:
    properties:
      createdAt:
        type: string
      employeeId:
        type: string
      entryType:
        type: string
      id:
        type: string
      quantity:
        type: number
      status:
        type: string
      updatedAt:
        type: string
      workDate:
        type: string
    type: object
  hrms_modules_worklog_internal_dto.PTItem:
    properties:
      createdAt:
        type: string
      employeeId:
        type: string
      eveningIn:
        type: string
      eveningMinutes:
        type: integer
      eveningOut:
        type: string
      id:
        type: string
      morningIn:
        type: string
      morningMinutes:
        type: integer
      morningOut:
        type: string
      status:
        type: string
      totalHours:
        type: number
      totalMinutes:
        type: integer
      updatedAt:
        type: string
      workDate:
        type: string
    type: object
  hrms_modules_worklog_internal_feature_ft.CreateRequest:
    properties:
      employeeId:
        type: string
      entryType:
        enum:
        - late
        - leave_day
        - leave_double
        - leave_hours
        - ot
        type: string
      quantity:
        type: number
      workDate:
        type: string
    required:
    - employeeId
    - entryType
    - quantity
    - workDate
    type: object
  hrms_modules_worklog_internal_feature_ft.CreateResponse:
    properties:
      createdAt:
        type: string
      employeeId:
        type: string
      entryType:
        type: string
      id:
        type: string
      quantity:
        type: number
      status:
        type: string
      updatedAt:
        type: string
      workDate:
        type: string
    type: object
  hrms_modules_worklog_internal_feature_ft.GetResponse:
    properties:
      createdAt:
        type: string
      employeeId:
        type: string
      entryType:
        type: string
      id:
        type: string
      quantity:
        type: number
      status:
        type: string
      updatedAt:
        type: string
      workDate:
        type: string
    type: object
  hrms_modules_worklog_internal_feature_ft.ListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/hrms_modules_worklog_internal_dto.FTItem'
        type: array
      meta:
        properties:
          currentPage:
            type: integer
          totalItems:
            type: integer
          totalPages:
            type: integer
        type: object
    type: object
  hrms_modules_worklog_internal_feature_ft.UpdateRequest:
    properties:
      entryType:
        enum:
        - late
        - leave_day
        - leave_double
        - leave_hours
        - ot
        type: string
      quantity:
        type: number
      status:
        enum:
        - pending
        - approved
        type: string
      workDate:
        type: string
    type: object
  hrms_modules_worklog_internal_feature_ft.UpdateResponse:
    properties:
      createdAt:
        type: string
      employeeId:
        type: string
      entryType:
        type: string
      id:
        type: string
      quantity:
        type: number
      status:
        type: string
      updatedAt:
        type: string
      workDate:
        type: string
    type: object
  hrms_modules_worklog_internal_feature_pt.CreateRequest:
    properties:
      employeeId:
        type: string
      eveningIn:
        type: string
      eveningOut:
        type: string
      morningIn:
        type: string
      morningOut:
        type: string
      status:
        enum:
        - pending
        - approved
        type: string
      workDate:
        type: string
    required:
    - employeeId
    - workDate
    type: object
  hrms_modules_worklog_internal_feature_pt.CreateResponse:
    properties:
      createdAt:
        type: string
      employeeId:
        type: string
      eveningIn:
        type: string
      eveningMinutes:
        type: integer
      eveningOut:
        type: string
      id:
        type: string
      morningIn:
        type: string
      morningMinutes:
        type: integer
      morningOut:
        type: string
      status:
        type: string
      totalHours:
        type: number
      totalMinutes:
        type: integer
      updatedAt:
        type: string
      workDate:
        type: string
    type: object
  hrms_modules_worklog_internal_feature_pt.GetResponse:
    properties:
      createdAt:
        type: string
      employeeId:
        type: string
      eveningIn:
        type: string
      eveningMinutes:
        type: integer
      eveningOut:
        type: string
      id:
        type: string
      morningIn:
        type: string
      morningMinutes:
        type: integer
      morningOut:
        type: string
      status:
        type: string
      totalHours:
        type: number
      totalMinutes:
        type: integer
      updatedAt:
        type: string
      workDate:
        type: string
    type: object
  hrms_modules_worklog_internal_feature_pt.ListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/hrms_modules_worklog_internal_dto.PTItem'
        type: array
      meta:
        properties:
          currentPage:
            type: integer
          totalItems:
            type: integer
          totalPages:
            type: integer
        type: object
    type: object
  hrms_modules_worklog_internal_feature_pt.UpdateRequest:
    properties:
      eveningIn:
        type: string
      eveningOut:
        type: string
      morningIn:
        type: string
      morningOut:
        type: string
      status:
        enum:
        - pending
        - approved
        type: string
      workDate:
        type: string
    type: object
  hrms_modules_worklog_internal_feature_pt.UpdateResponse:
    properties:
      createdAt:
        type: string
      employeeId:
        type: string
      eveningIn:
        type: string
      eveningMinutes:
        type: integer
      eveningOut:
        type: string
      id:
        type: string
      morningIn:
        type: string
      morningMinutes:
        type: integer
      morningOut:
        type: string
      status:
        type: string
      totalHours:
        type: number
      totalMinutes:
        type: integer
      updatedAt:
        type: string
      workDate:
        type: string
    type: object
...
info:
  contact: {}
  description: HR Payroll REST API
  title: HR Payroll API
  version: "1.0"
paths:
...
  /worklogs/ft:
    get:
      description: รายการ worklog (Full-time)
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: employee id
        in: query
        name: employeeId
        type: string
      - description: pending|approved|all
        in: query
        name: status
        type: string
      - description: late|leave_day|leave_double|leave_hours|ot
        in: query
        name: entryType
        type: string
      - description: YYYY-MM-DD
        in: query
        name: startDate
        type: string
      - description: YYYY-MM-DD
        in: query
        name: endDate
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hrms_modules_worklog_internal_feature_ft.ListResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
      - BearerAuth: []
      summary: List worklogs FT
      tags:
      - Worklogs FT
    post:
      consumes:
      - application/json
      description: บันทึก worklog (Full-time) สถานะเริ่มต้น pending
      parameters:
      - description: worklog payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hrms_modules_worklog_internal_feature_ft.CreateRequest'
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/hrms_modules_worklog_internal_feature_ft.CreateResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "409":
          description: Conflict
      security:
      - BearerAuth: []
      summary: Create worklog FT
      tags:
      - Worklogs FT
  /worklogs/ft/{id}:
    delete:
      description: ลบ worklog (Full-time) ได้เฉพาะสถานะ pending
      parameters:
      - description: worklog id
        in: path
        name: id
        required: true
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Delete worklog FT
      tags:
      - Worklogs FT
    get:
      description: ดึง worklog (Full-time) ตาม id
      parameters:
      - description: worklog id
        in: path
        name: id
        required: true
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hrms_modules_worklog_internal_feature_ft.GetResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Get worklog FT detail
      tags:
      - Worklogs FT
    patch:
      consumes:
      - application/json
      description: แก้ไข worklog (Full-time). เปลี่ยนสถานะเป็น approved ได้, revert
        approved ไม่ได้
      parameters:
      - description: worklog id
        in: path
        name: id
        required: true
        type: string
      - description: worklog payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hrms_modules_worklog_internal_feature_ft.UpdateRequest'
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hrms_modules_worklog_internal_feature_ft.UpdateResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "409":
          description: Conflict
      security:
      - BearerAuth: []
      summary: Update worklog FT
      tags:
      - Worklogs FT
  /worklogs/pt:
    get:
      description: รายการ worklog (Part-time)
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: employee id
        in: query
        name: employeeId
        type: string
      - description: pending|approved|all
        in: query
        name: status
        type: string
      - description: YYYY-MM-DD
        in: query
        name: startDate
        type: string
      - description: YYYY-MM-DD
        in: query
        name: endDate
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hrms_modules_worklog_internal_feature_pt.ListResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
      - BearerAuth: []
      summary: List worklogs PT
      tags:
      - Worklogs PT
    post:
      consumes:
      - application/json
      description: บันทึก worklog (Part-time)
      parameters:
      - description: worklog payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hrms_modules_worklog_internal_feature_pt.CreateRequest'
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/hrms_modules_worklog_internal_feature_pt.CreateResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
      - BearerAuth: []
      summary: Create worklog PT
      tags:
      - Worklogs PT
  /worklogs/pt/{id}:
    delete:
      description: ลบ worklog (Part-time) ได้เฉพาะสถานะ pending
      parameters:
      - description: worklog id
        in: path
        name: id
        required: true
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Delete worklog PT
      tags:
      - Worklogs PT
    get:
      description: ดึง worklog (Part-time) ตาม id
      parameters:
      - description: worklog id
        in: path
        name: id
        required: true
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hrms_modules_worklog_internal_feature_pt.GetResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Get worklog PT detail
      tags:
      - Worklogs PT
    patch:
      consumes:
      - application/json
      description: แก้ไข worklog (Part-time). เปลี่ยนสถานะเป็น approved ได้, revert
        approved ไม่ได้
      parameters:
      - description: worklog id
        in: path
        name: id
        required: true
        type: string
      - description: worklog payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hrms_modules_worklog_internal_feature_pt.UpdateRequest'
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hrms_modules_worklog_internal_feature_pt.UpdateResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Update worklog PT
      tags:
      - Worklogs PT
securityDefinitions:
  BearerAuth:
    description: Add "Bearer <access token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"